Take two at rewriting menu options
