+++ src/Widgets.h
@@ -28,9 +28,6 @@ CC_NOINLINE void TextWidget_Set(struct TextWidget* w, const cc_string* text, str
 /* Shorthand for TextWidget_Set using String_FromReadonly */
 CC_NOINLINE void TextWidget_SetConst(struct TextWidget* w, const char* text, struct FontDesc* font);
 
-
-typedef void (*Button_Get)(cc_string* raw);
-typedef void (*Button_Set)(const cc_string* raw);
 /* A labelled button that can be clicked on. */
 struct ButtonWidget {
 	Widget_Body
@@ -38,8 +35,6 @@ struct ButtonWidget {
 	PackedCol color;
 	int minWidth, minHeight;
 	const char* optName;
-	Button_Get GetValue;
-	Button_Set SetValue;
 };
 #define BUTTONWIDGET_MAX 12
